requirements.txt
Appium-Python-Client==0.50
selenium==3.141.0
pytest==5.4.1
